fix: resolve file_length lint violations from the weekly reset guard
Both touched files sat exactly at SwiftLint's 1500-line cap; the previous
commit's additions pushed each over it.

- Consolidate the Codex and Claude weekly branches in
  postLimitResetCelebrationIfNeeded into one, since they now differ only in
  requiresPreviousBoundary. Net code-line count is unchanged from main.
- Move the new Claude weekly transient-zero test into its own file,
  UsageStorePlanUtilizationClaudeWeeklyResetGuardTests.swift, matching the
  existing UsageStorePlanUtilizationCodexResetOwnershipTests.swift split.
